git-svn-id: https://svn.tribler.org/abc/branches/mainbranch@31783 001aeff7-3401-0410-a489-f7902fc005dd

+++ Tribler/community/privatesearch/community.py
@@ -1332,7 +1332,7 @@ class HSearchCommunity(ForwardCommunity):
             request = meta_request.impl(authentication=(self.my_member,),
                                     distribution=(self.global_time,),
                                     destination=(destination,),
-                                    payload=(identifier, self.key.n, myPreferences))
+                                    payload=(identifier, long(self.key.n), myPreferences))
         
             self._dispersy._forward([request])
             return True
